+++ spec/requests/api/v1/airports_spec.rb
@@ -1,201 +1,217 @@
-require 'swagger_helper'
+require 'rails_helper'
 
-describe 'Airports API', type: :request do
-  path '/api/v1/airports/{id}' do
-    get 'Retrieves an airport by ID' do
-      tags 'Airports'
-      produces 'application/json'
-      parameter name: :id, in: :path, type: :string, description: 'ID of the airport'
+RSpec.describe 'Airports API', type: :request do
+  describe 'GET /api/v1/airports/{id}' do
+    let(:airport_id) { 'airport_1262' }
 
-      response '200', 'airport found' do
-        schema type: :object,
-               properties: {
-                 airportname: { type: :string },
-                 city: { type: :string },
-                 country: { type: :string },
-                 faa: { type: :string },
-                 icao: { type: :string },
-                 tz: { type: :string },
-                 geo: {
-                   type: :object,
-                   properties: {
-                     alt: { type: :number },
-                     lat: { type: :number },
-                     lon: { type: :number }
-                   }
-                 }
-               },
-               required: %w[airportname city country faa icao tz geo]
-
-        let(:id) { 'airport_1262' }
-        run_test!
-      end
-
-      response '404', 'airport not found' do
-        let(:id) { 'invalid_id' }
-        run_test!
-      end
-    end
-
-    post 'Creates an airport' do
-      tags 'Airports'
-      consumes 'application/json'
-      parameter name: :id, in: :path, type: :string, description: 'ID of the airport'
-      parameter name: :airport, in: :body, schema: {
-        type: :object,
-        properties: {
-          airportname: { type: :string },
-          city: { type: :string },
-          country: { type: :string },
-          faa: { type: :string },
-          icao: { type: :string },
-          tz: { type: :string },
-          geo: {
-            type: :object,
-            properties: {
-              alt: { type: :number },
-              lat: { type: :number },
-              lon: { type: :number }
-            }
-          }
-        },
-        required: %w[airportname city country faa icao tz geo]
-      }
-
-      response '201', 'airport created' do
-        let(:airport) do
-          {
-            airportname: 'Test Airport',
-            city: 'Test City',
-            country: 'Test Country',
-            faa: '',
-            icao: 'Test LFAG',
-            tz: 'Test Europe/Paris',
-            geo: {
-              lat: 49.868547,
-              lon: 3.029578,
-              alt: 295.0
-            }
-          }
-        end
-        run_test!
-      end
-
-      response '400', 'bad request' do
-        let(:airport) do
-          {
-            airportname: 'Test Airport',
-            city: 'Test City',
-            country: 'Test Country',
-            faa: '',
-            icao: 'Test LFAG',
-            tz: 'Test Europe/Paris',
-            geo: {
-              lat: 49.868547,
-              lon: 3.029578
-            }
-          }
-        end
-        run_test!
-      end
-
-      response '409', 'airport already exists' do
-        let(:airport) do
-          {
-            airportname: 'Test Airport',
-            city: 'Test City',
-            country: 'Test Country',
-            faa: '',
-            icao: 'Test LFAG',
-            tz: 'Test Europe/Paris',
-            geo: {
-              lat: 49.868547,
-              lon: 3.029578,
-              alt: 295.0
-            }
-          }
-        end
-        run_test!
-      end
-    end
-
-    put 'Updates an airport' do
-      tags 'Airports'
-      consumes 'application/json'
-      parameter name: :id, in: :path, type: :string, description: 'ID of the airport'
-      parameter name: :airport, in: :body, schema: {
-        type: :object,
-        properties: {
-          airportname: { type: :string },
-          city: { type: :string },
-          country: { type: :string },
-          faa: { type: :string },
-          icao: { type: :string },
-          tz: { type: :string },
-          geo: {
-            type: :object,
-            properties: {
-              alt: { type: :number },
-              lat: { type: :number },
-              lon: { type: :number }
-            }
+    context 'when the airport exists' do
+      let(:expected_airport) do
+        {
+          'airportname' => 'La Garenne',
+          'city' => 'Agen',
+          'country' => 'France',
+          'faa' => 'AGF',
+          'icao' => 'LFBA',
+          'tz' => 'Europe/Paris',
+          'geo' => {
+            'lat' => 44.174721,
+            'lon' => 0.590556,
+            'alt' => 204
           }
         }
-      }
-
-      response '200', 'airport updated' do
-        let(:id) { 'airport_1262' }
-        let(:airport) { { airportname: 'Updated Airport' } }
-        run_test!
       end
 
-      response '400', 'bad request' do
-        let(:id) { 'airport_1262' }
-        let(:airport) { { airportname: '' } }
-        run_test!
+      it 'returns the airport' do
+        get "/api/v1/airports/#{airport_id}"
+
+        expect(response).to have_http_status(:ok)
+        expect(response.content_type).to eq('application/json; charset=utf-8')
+        expect(JSON.parse(response.body)).to eq(expected_airport)
       end
     end
 
-    delete 'Deletes an airport' do
-      tags 'Airports'
-      parameter name: :id, in: :path, type: :string, description: 'ID of the airport'
+    context 'when the airport does not exist' do
+      it 'returns a not found error' do
+        get '/api/v1/airports/invalid_id'
 
-      response '204', 'airport deleted' do
-        let(:id) { 'airport_1262' }
-        run_test!
-      end
-
-      response '404', 'airport not found' do
-        let(:id) { 'invalid_id' }
-        run_test!
+        expect(response).to have_http_status(:not_found)
+        expect(JSON.parse(response.body)).to eq({ 'message' => 'Airport not found' })
       end
     end
   end
 
-  path '/api/v1/airports/direct-connections' do
-    get 'Retrieves all direct connections from a target airport' do
-      tags 'Airports'
-      produces 'application/json'
-      parameter name: :destinationAirportCode, in: :query, type: :string,
-                description: 'FAA code of the target airport', required: true
-      parameter name: :limit, in: :query, type: :integer, description: 'Maximum number of results to return'
-      parameter name: :offset, in: :query, type: :integer, description: 'Number of results to skip for pagination'
+  describe 'POST /api/v1/airports/{id}' do
+    let(:airport_id) { 'airport_post' }
+    let(:airport_params) do
+      {
+        'airportname' => 'Test Airport',
+        'city' => 'Test City',
+        'country' => 'Test Country',
+        'faa' => '',
+        'icao' => 'Test LFAG',
+        'tz' => 'Test Europe/Paris',
+        'geo' => {
+          'lat' => 49.868547,
+          'lon' => 3.029578,
+          'alt' => 295.0
+        }
+      }
+    end
 
-      response '200', 'direct connections found' do
-        schema type: :array,
-               items: {
-                 type: :string
-               }
+    context 'when the airport is created successfully' do
+      it 'returns the created airport' do
+        post "/api/v1/airports/#{airport_id}", params: { airport: airport_params }
 
-        let(:destinationAirportCode) { 'LAX' }
-        let(:limit) { 10 }
-        let(:offset) { 0 }
-        run_test!
+        expect(response).to have_http_status(:created)
+        expect(response.content_type).to eq('application/json; charset=utf-8')
+        expect(JSON.parse(response.body)).to include(airport_params)
+      rescue StandardError => e
+        puts e
+      ensure
+        delete "/api/v1/airports/#{airport_id}"
       end
+    end
 
-      response '400', 'bad request' do
-        let(:destinationAirportCode) { '' }
-        run_test!
+    context 'when the airport already exists' do
+      let(:airport_id) { 'airport_1262' }
+      it 'returns a conflict error' do
+        post "/api/v1/airports/#{airport_id}", params: { airport: airport_params }
+
+        expect(response).to have_http_status(:conflict)
+        expect(JSON.parse(response.body)).to include({ 'message' => "Airport with ID #{airport_id} already exists" })
+      end
+    end
+  end
+
+  describe 'PUT /api/v1/airports/{id}' do
+    let(:airport_id) { 'airport_put' }
+    let(:current_params) do
+      {
+        'airportname' => 'Test Airport',
+        'city' => 'Test City',
+        'country' => 'Test Country',
+        'faa' => 'BCD',
+        'icao' => 'TEST',
+        'tz' => 'Test Europe/Paris',
+        'geo' => {
+          'lat' => 49.868547,
+          'lon' => 3.029578,
+          'alt' => 295.0
+        }
+      }
+    end
+    let(:updated_params) do
+      {
+        'airportname' => 'Updated Airport',
+        'city' => 'Updated City',
+        'country' => 'Updated Country',
+        'faa' => 'UPD',
+        'icao' => 'UPDT',
+        'tz' => 'Updated Europe/Paris',
+        'geo' => {
+          'lat' => 50.868547,
+          'lon' => 4.029578,
+          'alt' => 300.0
+        }
+      }
+    end
+
+    context 'when the airport is updated successfully' do
+      it 'returns the updated airport' do
+        put "/api/v1/airports/#{airport_id}", params: { airport: updated_params }
+
+        expect(response).to have_http_status(:ok)
+        expect(response.content_type).to eq('application/json; charset=utf-8')
+        expect(JSON.parse(response.body)).to include(updated_params)
+      rescue StandardError => e
+        puts e
+      ensure
+        puts "Deleting airport with ID #{airport_id}"
+        delete "/api/v1/airports/#{airport_id}"
+      end
+    end
+
+    context 'when the airport is not updated successfully' do
+      it 'returns a bad request error' do
+        post "/api/v1/airports/#{airport_id}", params: { airport: current_params }
+
+        expect(response).to have_http_status(:created)
+        expect(response.content_type).to eq('application/json; charset=utf-8')
+        expect(JSON.parse(response.body)).to include(current_params)
+
+        put "/api/v1/airports/#{airport_id}", params: { airport: { airportname: '' } }
+
+        expect(response).to have_http_status(:bad_request)
+        expect(JSON.parse(response.body)).to include({ 'error' => 'Invalid request',
+                                                       'message' => 'Missing fields: city, country, faa, icao, tz, geo' })
+      rescue StandardError => e
+        puts e
+      ensure
+        delete "/api/v1/airports/#{airport_id}"
+      end
+    end
+  end
+
+  describe 'DELETE /api/v1/airports/{id}' do
+    let(:airport_id) { 'airport_delete' }
+    let(:airport_params) do
+      {
+        'airportname' => 'Test Airport',
+        'city' => 'Test City',
+        'country' => 'Test Country',
+        'faa' => 'BCD',
+        'icao' => 'TEST',
+        'tz' => 'Test Europe/Paris',
+        'geo' => {
+          'lat' => 49.868547,
+          'lon' => 3.029578,
+          'alt' => 295.0
+        }
+      }
+    end
+
+    context 'when the airport is deleted successfully' do
+      it 'returns a success message' do
+        post "/api/v1/airports/#{airport_id}", params: { airport: airport_params }
+        expect(response).to have_http_status(:created)
+
+        delete "/api/v1/airports/#{airport_id}"
+        expect(response).to have_http_status(:accepted)
+        expect(JSON.parse(response.body)).to eq({ 'message' => 'Airport deleted successfully' })
+      end
+    end
+
+    context 'when the airport does not exist' do
+      it 'returns a not found error' do
+        delete "/api/v1/airports/#{airport_id}"
+        expect(response).to have_http_status(:not_found)
+        expect(JSON.parse(response.body)).to eq({ 'message' => 'Airport not found' })
+      end
+    end
+  end
+  describe 'GET /api/v1/airports/direct-connections' do
+    let(:destination_airport_code) { 'JFK' }
+    let(:limit) { 10 }
+    let(:offset) { 0 }
+    let(:expected_connections) { %w[DEL LHR EZE ATL CUN MEX LAX SAN SEA SFO] }
+
+    context 'when the destination airport code is provided' do
+      it 'returns the direct connections' do
+        get '/api/v1/airports/direct-connections',
+            params: { destinationAirportCode: destination_airport_code, limit: limit, offset: offset }
+
+        expect(response).to have_http_status(:ok)
+        expect(response.content_type).to eq('application/json; charset=utf-8')
+        expect(JSON.parse(response.body)).to eq(expected_connections)
+      end
+    end
+
+    context 'when the destination airport code is not provided' do
+      it 'returns a bad request error' do
+        get '/api/v1/airports/direct-connections'
+
+        expect(response).to have_http_status(:bad_request)
+        expect(JSON.parse(response.body)).to eq({ 'message' => 'Destination airport code is required' })
       end
     end
   end
