+++ spec/requests/swagger/hotels_spec.rb
